% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_distinct_col_count_prop.R
\name{group_distinct_col_count_prop}
\alias{group_distinct_col_count_prop}
\title{Group by a Column and Calculate Proportion of Distinct Values in Another Column}
\usage{
group_distinct_col_count_prop(dataset, group_col, summary_col)
}
\arguments{
\item{dataset}{A data frame containing the data to be grouped and summarized.}

\item{group_col}{A string representing the name of the column to group by.}

\item{summary_col}{A string representing the name of the column to calculate
the proportion of distinct values.}
}
\value{
: A data frame containing the grouping column, the total count of
distinct values in the summary_col for each group, and the proportion of
these distinct values relative to the total distinct values in
summary_col across the entire dataset.
}
\description{
This function groups a given dataset by a specified column and then calculates
the proportion of distinct values in another specified column within each
group. It is designed to provide a quick and generic way to summarize data,
which is useful in exploratory data analysis.
}
\examples{
library(dplyr)

# Example data frame
test_data <- data.frame(
  Species = c("Oak", "Maple", "Oak", "Pine", "Maple", "Pine", NA),
  Height = c(10, 15, 20, 25, 30, 35, 40)
)

# Usage of the group_distinct_col_count_prop function for with no NA's
result1 <- group_distinct_col_count_prop(test_data[1:6, ], "Species", "Height")

# Usage of the group_distinct_col_count_prop function for with NA's
result2 <- group_distinct_col_count_prop(test_data, "Species", "Height")

}
